+++ CHANGELOG.md
@@ -10,10 +10,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and 
 ### Changed
 
 - The plugin now requires:
-	- `stylelint` version `17.0.0` or higher,
+	- `stylelint` version `17.1.0` or higher,
 	- `node.js` version `20.19.0` or higher.
 - Hex colors are no longer allowed.
 - Color function names with alpha are no longer allowed.
+
+### Fixed
+
+- The `display` property now not only disallows legacy and short notation, but also automatically fixes it to the full modern notation.
 
 ### Added